Move the generation of the central directory entry to the ZipInfo object.
Moving this logic places it closer to where the local header is
generated and closer to other logic related to the information written
to the central directory.

The option of using a separate `encode_central_directory` method to
construct bytes separate from the main `central_directory` method is to
allow allows sub classes to alter fields before they are written out,
without having to copy the rest of the `central_directory` method.