Change local and central directory encoding to always use a new extra block
ZipInfo.extra is populated with the extra from the central directory
when reading from an existing zipfile. When using a ZipInfo read from
the central directory as the basis of a new file entry, the contents of
the existing ZipInfo.extra bytes could be written to the new file.

I was a bit uncertain about the safety / accuracy of reusing extra block
fields that we hadn't cleaned or verified. Add to it that we may change
something which would invalidate one of the extra block regions that we
copied across without updating it.

For instance, if the original file had a zip64 extra block but the new
file didn't require a zip64 extra block, it was possible that the old
zip64 extra block would have been copied across anyway.

Also the new local directory would use the same ZipInfo.extra which was
read from the central directory. It is possible for different fields to
be present in the central and local directories (zip64 does this for
instance). This could have resulted in an incorrect field.

While these fields may have just been ignored, I think it is better to
start a fresh extra block every time.