Update decode_extra_zip64 to have more information about the spec
This change is mostly cosmetic to illustrate some aspects in the pkware
spec that we don't (currently) care about.

Theoretically, it would now be able to read the zip64 extra block in a
central directory of a zipfile which has > 65,535 parts... but as
zipfile doesn't handle multipart files, it is a moot point.