Change raised exception when writing unexpected zip64 stream
A struct exception for trying to write an unsigned long long into an
unsigned long was raised when attempting to write the data descriptor
for a zip64 sized stream when zip64 wasn't enabled.

It now raises the same exception as when this happens and the data
descriptor is not being used.